requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pyodbc
azure.storage.blob
numpy
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.7.0+cpu 
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.8.1+cpu 
-f https://download.pytorch.org/whl/torch_stable.html
torchaudio==0.7.0
attacut